LPS-49191 new namespace pattern to match the *Portlet and *Upgrade classes
