+++ modules/apps/social/social-activities-web/src/com/liferay/social/activities/portlet/action/SocialActivitiesConfigurationAction.java
@@ -29,5 +29,6 @@ import org.osgi.service.component.annotations.Component;
 	},
 	service = ConfigurationAction.class
 )
-public class ConfigurationActionImpl extends DefaultConfigurationAction {
+public class SocialActivitiesConfigurationAction
+	extends DefaultConfigurationAction {
 }
